Fix for non-compiling NL code (hopefully)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2236 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ libs/hawknl/src/err.c
@@ -110,7 +110,7 @@ NL_EXP NLenum NL_APIENTRY nlGetError(void)
     }
     else
     {
-        result = (NLenum)pthread_getspecific(key);
+        result = (NLenum)((long)pthread_getspecific(key));
         return result;
     }
 #endif
